vastly simplify RHS state
