+++ webapp/src/components/rhs/rhs_main.tsx
@@ -1,45 +1,25 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import React, {FC, useEffect} from 'react';
+import React, {FC, useEffect, useState} from 'react';
 import {useDispatch, useSelector} from 'react-redux';
-import styled from 'styled-components';
 
 import {GlobalState} from 'mattermost-redux/types/store';
+import {getCurrentChannelId} from 'mattermost-redux/selectors/entities/channels';
 
-import {setRHSOpen} from 'src/actions';
-import Spinner from 'src/components/assets/icons/spinner';
-import RHSIncidentDetails from 'src/components/rhs/incident_details';
+import {setRHSOpen, setRHSState} from 'src/actions';
 import RHSListView from 'src/components/rhs/rhs_list_view';
-import {RHSContainer, RHSContent} from 'src/components/rhs/rhs_shared';
-import {IncidentFetchState, useCurrentIncident} from 'src/hooks';
-import {currentRHSState} from 'src/selectors';
+import {currentRHSState, isIncidentChannel} from 'src/selectors';
 import {RHSState} from 'src/types/rhs';
 import RHSWelcomeView from 'src/components/rhs/rhs_welcome_view';
-
-export const SpinnerContainer = styled.div`
-    text-align: center;
-    padding: 20px;
-`;
-
-const spinner = (
-    <RHSContainer>
-        <RHSContent>
-            <SpinnerContainer>
-                <Spinner/>
-                <span>{'Loading...'}</span>
-            </SpinnerContainer>
-        </RHSContent>
-    </RHSContainer>
-);
+import RHSDetailsView from 'src/components/rhs/rhs_details_view';
 
 const RightHandSidebar: FC<null> = () => {
     const dispatch = useDispatch();
+    const currentChannelId = useSelector<GlobalState, string>(getCurrentChannelId);
+    const inIncidentChannel = useSelector<GlobalState, boolean>((state) => isIncidentChannel(state, currentChannelId));
     const rhsState = useSelector<GlobalState, RHSState>(currentRHSState);
-
-    // Only get the current incident, and incidentList at the top of the rhs hierarchy.
-    // This prevents race conditions.
-    const [incident, incidentFetchState] = useCurrentIncident();
+    const [seenChannelId, setSeenChannelId] = useState('');
 
     useEffect(() => {
         dispatch(setRHSOpen(true));
@@ -48,39 +28,25 @@ const RightHandSidebar: FC<null> = () => {
         };
     }, [dispatch]);
 
-    if (rhsState === RHSState.WelcomeScreen) {
+    // Update the rhs state when the channel changes
+    if (currentChannelId !== seenChannelId) {
+        setSeenChannelId(currentChannelId);
+
+        if (inIncidentChannel) {
+            dispatch(setRHSState(RHSState.ViewingIncident));
+        } else {
+            dispatch(setRHSState(RHSState.ViewingList));
+        }
+    }
+
+    if (rhsState === RHSState.ViewingIncident) {
+        if (inIncidentChannel) {
+            return <RHSDetailsView/>;
+        }
         return <RHSWelcomeView/>;
     }
 
-    if (rhsState === RHSState.ViewingList) {
-        return (
-            <RHSContainer>
-                <RHSContent>
-                    <RHSListView
-                        currentIncidentId={incident?.id}
-                    />
-                </RHSContent>
-            </RHSContainer>
-        );
-    }
-
-    if (incidentFetchState === IncidentFetchState.Loading) {
-        return spinner;
-    } else if (incidentFetchState === IncidentFetchState.NotFound || incident === null) {
-        // This should not happen--if incident is not found or null, we should be viewing the list.
-        // Returning the spinner so that if it ever happens, we at least show something.
-        return spinner;
-    }
-
-    return (
-        <RHSContainer>
-            <RHSContent>
-                <RHSIncidentDetails
-                    incident={incident}
-                />
-            </RHSContent>
-        </RHSContainer>
-    );
+    return <RHSListView/>;
 };
 
 export default RightHandSidebar;
