+++ webapp/src/types/actions.ts
@@ -13,7 +13,6 @@ export const SET_CLIENT_ID = pluginId + '_set_client_id';
 export const INCIDENT_CREATED = pluginId + '_incident_created';
 export const INCIDENT_UPDATED = pluginId + '_incident_updated';
 export const RECEIVED_TEAM_INCIDENTS = pluginId + '_received_team_incident_channels';
-export const FETCHING_INCIDENTS = pluginId + '_fetching_incidents';
 export const REMOVED_FROM_INCIDENT_CHANNEL = pluginId + '_removed_from_incident_channel';
 export const SET_RHS_STATE = pluginId + '_set_rhs_state';
 
@@ -50,11 +49,6 @@ export interface IncidentUpdated {
 export interface ReceivedTeamIncidents {
     type: typeof RECEIVED_TEAM_INCIDENTS;
     incidents: Incident[];
-}
-
-export interface FetchingIncidents {
-    type: typeof FETCHING_INCIDENTS;
-    fetching: boolean;
 }
 
 export interface RemovedFromIncidentChannel {
