+++ webapp/src/types/rhs.ts
@@ -4,5 +4,4 @@
 export enum RHSState {
     ViewingList,
     ViewingIncident,
-    WelcomeScreen,
 }
